Feature: Add dynamic worker pool

+++ internal/app/usecases/downloader/internal/file_status/status_to_working.go
@@ -12,7 +12,7 @@ func (uc *FileStatus) Working(
 	ctx context.Context,
 	fileId uuid.UUID,
 	taskId uuid.UUID,
-	workerId uint,
+	workerId uint64,
 ) error {
 	err := uc.dlTaskStatus.Working(ctx, taskId, workerId)
 	if err != nil {
